move expire calculation from supergraph_service to authentication mod
As per code review suggestion.
